third_party/blink/web_tests/wpt_internal/css/css-overflow/column-scroll-marker-009.html

<!DOCTYPE html>
<title>CSS Overflow Test: ::column::scroll-marker supports :focus</title>
<link rel="help" href="https://github.com/flackr/carousel/tree/main/fragmentation">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<style>
  #container {
    overflow: hidden;
    columns: 3;
    column-fill: auto;
    gap: 0;
    orphans: 1;
    widows: 1;
    height: 100px;
    border: 15px solid;
    line-height: 20px;
    scroll-marker-group: before;
    background: yellow;
  }
  #container::scroll-marker-group {
    display: flex;
    justify-content: space-between;
    height: 50px;
    background: cyan;
  }
  #container::column::scroll-marker {
    display: flex;
    justify-content: center;
    align-items: center;
    width: 50px;
    height: 50px;
    background: hotpink;
    content: "*";
  }
  #container::column::scroll-marker:focus {
    background: blue;
  }
</style>
<div style="width:450px;">
  <div id="container">
    First column<br>
    <br>
    <br>
    <br>
    <br>
    Second column<br>
    <br>
    <br>
    <br>
    <br>
    Third column<br>
    <br>
    <br>
    <br>
    <br>
    Fourth column<br>
    <br>
    <br>
    <br>
    <br>
    Fifth column<br>
  </div>
</div>
<script>
  promise_test(async t => {
    actions_promise = new test_driver.Actions()
    .pointerMove(15, 15)
    .pointerDown()
    .pointerUp()
    .send();
    await actions_promise;
    assert_equals(getComputedStyle(container, "::column::scroll-marker").backgroundColor, "blue");
  });
</script>
